fix: synchronize entities with database schema to resolve SQL errors
Fixed multiple entity-database schema mismatches causing SQLGrammarException errors:

1. RuleBundleEntity:
   - Removed invalid 'subject' relationship (subject_id column doesn't exist in DB)
   - Added subjectType and subjectKey fields matching DB columns
   - Removed BaseEntity inheritance (DB only has id and created_at, not updated_at/created_by/updated_by/version)
   - Added manual ID generation with UUIDv7
   - Fixed index definition to match actual DB columns

2. RuleAssignmentEntity:
   - Removed invalid 'subject' relationship
   - Added subjectType and subjectKey fields matching DB columns
   - Updated index definitions
   - Updated constructor parameters

3. RuleAssignmentSubjectEntity:
   - Fixed column mapping: type -> subject_type
   - Added ruleAssignmentId field
   - Removed invalid OneToMany relationships to assignments and bundles
   - Updated index definition

4. Repository fixes:
   - RuleBundleRepository: Updated queries to use subjectType/subjectKey instead of subject.type/subject.key
   - RuleAssignmentRepository: Updated queries and added new finder methods

5. BundleHashResponse:
   - Changed ruleVersion type from Integer to Long to match DB bigint type

All entities now accurately match the Liquibase schema definitions.

+++ src/main/java/vn/viettel/vds/promotion/validation/adapter/in/web/dto/BundleHashResponse.java
@@ -26,7 +26,7 @@ public record BundleHashResponse(
         String bundleHash,
 
         @Schema(description = "Rule version number", example = "3")
-        Integer ruleVersion,
+        Long ruleVersion,
 
         @Schema(description = "Assignment version number", example = "1")
         Integer assignmentVersion,
